portreto/auth/django/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
idna==2.8
pycrypto==2.6.1
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.2

